Move GameTest to control/engine test package and rename class to EngineTest * Create evaluate and countMaterialValues methods in Engine class
